finish refactor for gist block. relates to #504.
