+++ src/blocks/gist/controls.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import icons from './../../../utils/icons';
+import icons from './../../utils/icons';
 
 /**
  * WordPress dependencies
